python/pyproject.toml
[tool.poetry]
name = "visionlib"
version = "1.0.0"
description = ""
authors = ["flonix8 <flstanek@googlemail.com>"]
license = "AGPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
valkey = ">=6.0.0"
pybase64 = ">=1.3.0"
pydantic = ">=2.3.0"
pyyaml = ">=6.0.1"
pydantic-settings = ">=2.0.3"
pyturbojpeg = ">=1.7.2"
numpy = ">=1.24.0"


[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.1"
testcontainers = ">=4.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
